+++ dag/demography.yml
@@ -276,12 +276,12 @@ steps:
   ########################################################################
 
   # Wittgenstein Centre (Projections)
-  data://meadow/demography/2024-12-06/wittgenstein_human_capital:
+  data://meadow/demography/2024-12-06/wittgenstein_human_capital_proj:
     - snapshot://demography/2024-12-06/wittgenstein_human_capital.zip
-  data://garden/demography/2024-12-06/wittgenstein_human_capital:
-    - data://meadow/demography/2024-12-06/wittgenstein_human_capital
-  data://grapher/demography/2024-12-06/wittgenstein_human_capital:
-    - data://garden/demography/2024-12-06/wittgenstein_human_capital
+  data://garden/demography/2024-12-06/wittgenstein_human_capital_proj:
+    - data://meadow/demography/2024-12-06/wittgenstein_human_capital_proj
+  data://grapher/demography/2024-12-06/wittgenstein_human_capital_proj:
+    - data://garden/demography/2024-12-06/wittgenstein_human_capital_proj
 
   # Wittgenstein Centre (Historical)
   data://meadow/demography/2024-12-06/wittgenstein_human_capital_historical:
@@ -290,6 +290,8 @@ steps:
     - data://meadow/demography/2024-12-06/wittgenstein_human_capital_historical
 
   # Wittgenstein Centre (Projections + Historical)
-  data://garden/demography/2024-12-06/wittgenstein_human_capital_combined:
+  data://garden/demography/2024-12-06/wittgenstein_human_capital:
     - data://meadow/demography/2024-12-06/wittgenstein_human_capital_historical
-    - data://meadow/demography/2024-12-06/wittgenstein_human_capital
+    - data://meadow/demography/2024-12-06/wittgenstein_human_capital_proj
+  data://grapher/demography/2024-12-06/wittgenstein_human_capital:
+    - data://garden/demography/2024-12-06/wittgenstein_human_capital
